Add possibility to store the inverse of camera link in the global link
Now you can have a camera with a camera link (due to hand eye or workspace calibration) and still get data in the camera frame. The global link will be the inverse of the of the camera link and the point cloud will be transformed into the cameras optical frame.