ensenso_camera: Fix finding of first available camera serial
The "BySerialNo" node has been hidden and linked to the "Cameras"
node, which also includes the two nodes "ByEepromId" and
"BySerialNo". When iterating over the items in the "Cameras" node, it 
could occur that the function tried to access the inexistent item 
"Available". The problem probably did not occur before, because the 
cameras are listed alphabetically and real cameras with numeric serial 
numbers are listed before the two nodes "ByEepromId" and "BySerialNo",
letting the function find and return a camera serial without causing
an exception.

We now check whether the item is a camera by requiring the
"Status/Available" node to be existing.