ensenso_camera: Fix default target frame if link frame is given
When specifying a linkFrame but no targetFrame, the latter was not set
to a default value but remained an empty string. This ultimately led to
tf being unable to look up the transformation from the linkFrame to the
targetFrame.

Fix it by setting the targetFrame to the linkFrame in this special case.
The resulting point clouds and calibration pattern poses, which are in
the linkFrame already (transformed by the NxLib), are now in the
targetFrame as expected.