stereo_camera: Fix usage of automatically adjusted parameters
For example, if auto exposure was enabled, the value calculated by the
NxLib was not updated in the parameter set. This had the effect that the
original start value was re-used when capturing the next time, and it
took several frames again to obtain a good result.

Save the parameter set after capture to update the automatically
adjusted values.